doctrenderer: make sure printf always uses a format string
Passing the string that is to be printed as the format string could
lead to trouble when the string has sequences that would have meaning
in a format string. Using the `"%s"` format string makes sure the
string that is to be printed isn't interpreted as a format string.

This also makes compiling with `-Werror=format-security` happy.

Signed-off-by: Arnout Engelen <arnout@bzzt.net>